+++ app/Http/Requests/UserProfileUpdateRequest.php
@@ -25,48 +25,33 @@
 
 namespace Gamify\Http\Requests;
 
-use Gamify\Models\User;
-use Illuminate\Validation\Rule;
-
 class UserProfileUpdateRequest extends Request
 {
-    public function authorize(): bool
-    {
-        /** @var User $user */
-        $user = $this->route('username');
-
-        return $user->username == $this->user()->username;
-    }
-
     public function rules(): array
     {
         return [
-            'url' => [
-                'sometimes',
-                'url',
+            'bio' => [
+                'nullable',
+                'string',
             ],
             'date_of_birth' => [
-                'sometimes',
+                'nullable',
                 'date',
             ],
-            'gender' => [
-                'required',
-                Rule::in(['male', 'female', 'unspecified']),
-            ],
             'twitter' => [
-                'sometimes',
+                'nullable',
                 'url',
             ],
             'facebook' => [
-                'sometimes',
+                'nullable',
                 'url',
             ],
             'linkedin' => [
-                'sometimes',
+                'nullable',
                 'url',
             ],
             'github' => [
-                'sometimes',
+                'nullable',
                 'url',
             ],
         ];