+++ app/Models/Badge.php
@@ -33,6 +33,7 @@ use Illuminate\Database\Eloquent\Casts\Attribute;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
+use Illuminate\Support\Str;
 use Laracodes\Presenter\Traits\Presentable;
 use QCod\ImageUp\HasImageUploads;
 
@@ -78,7 +79,9 @@ class Badge extends Model
             'file_input' => 'image',
         ],
     ];
+
     protected string $presenter = BadgePresenter::class;
+
     protected $fillable = [
         'name',
         'description',
@@ -108,6 +111,11 @@ class Badge extends Model
         return $query
             ->active()
             ->hasAnyFlags('actuators', $actuators);
+    }
+
+    public function slug(): string
+    {
+        return Str::slug($this->name);
     }
 
     protected function image(): Attribute