+++ app/Presenters/UserPresenter.php
@@ -37,4 +37,21 @@ class UserPresenter extends Presenter
     {
         return $this->model->role->description;
     }
+
+    public function createdAt(): string
+    {
+        return $this->model->created_at->toFormattedDateString();
+    }
+
+    public function birthdate(): string
+    {
+        return $this->model->profile->date_of_birth?->toFormattedDateString()
+            ?? __('user/profile.date_of_birth_not_set');
+    }
+
+    public function bio(): string
+    {
+        return $this->model->profile->bio
+            ?? '';
+    }
 }