+++ resources/views/layouts/master.blade.php
@@ -16,11 +16,6 @@
     <link rel="stylesheet" type="text/css" href="{{ asset('vendor/AdminLTE/bootstrap/css/bootstrap.min.css') }}">
     <link rel="stylesheet" type="text/css"
           href="{{ asset('//maxcdn.bootstrapcdn.com/font-awesome/4.4.0/css/font-awesome.min.css') }}">
-    <!-- end: GLOBAL CSS -->
-    <!-- start: CSS REQUIRED FOR THIS PAGE ONLY -->
-    @stack('styles')
-    <!-- end: CSS REQUIRED FOR THIS PAGE ONLY -->
-    <!-- start: MAIN CSS -->
     <link rel="stylesheet" type="text/css" href="{{ asset('vendor/AdminLTE/css/AdminLTE.min.css') }}">
     <link rel="stylesheet" type="text/css" href="{{ asset('vendor/AdminLTE/css/skins/skin-blue.min.css') }}">
     <!-- HTML5 Shim and Respond.js IE8 support of HTML5 elements and media queries -->
@@ -29,7 +24,10 @@
     <script type="text/javascript" src="{{ asset('//oss.maxcdn.com/html5shiv/3.7.3/html5shiv.min.js') }}"></script>
     <script type="text/javascript" src="{{ asset('//oss.maxcdn.com/respond/1.4.2/respond.min.js') }}"></script>
     <![endif]-->
-    <!-- end: MAIN CSS -->
+    <!-- end: GLOBAL CSS -->
+    <!-- start: CSS REQUIRED FOR THIS PAGE ONLY -->
+    @stack('styles')
+    <!-- end: CSS REQUIRED FOR THIS PAGE ONLY -->
     <link rel="shortcut icon" href="{{ asset('favicon.ico') }}">
 </head>
 <!-- end: HEAD -->
@@ -75,13 +73,12 @@
 <!-- start: GLOBAL JAVASCRIPT -->
 <script src="{{ asset('vendor/AdminLTE/jquery/jquery.min.js') }}"></script>
 <script src="{{ asset('vendor/AdminLTE/bootstrap/js/bootstrap.min.js') }}"></script>
+<script src="{{ asset('vendor/AdminLTE/js/adminlte.min.js') }}"></script>
+<script src="{{ asset('js/app.js') }}"></script>
 <!-- end: GLOBAL JAVASCRIPT -->
 <!-- start: JAVASCRIPT REQUIRED FOR THIS PAGE ONLY -->
 @stack('scripts')
 <!-- end: JAVASCRIPT REQUIRED FOR THIS PAGE ONLY -->
-<!-- start: MAIN JAVASCRIPT -->
-<script src="{{ asset('vendor/AdminLTE/js/adminlte.min.js') }}"></script>
-<!-- end: MAIN JAVASCRIPT -->
 </body>
 <!-- end: BODY -->
 </html>