+++ resources/views/profile/_about_me.blade.php
@@ -5,7 +5,7 @@
     <!-- /.box-header -->
     <div class="box-body">
         <div class="text-center">
-            @if (!empty($user->profile->facebook))
+            @if ($user->profile->facebook)
                 <a class="btn btn-social-icon btn-facebook" href="{{ $user->profile->facebook }}"
                    rel="nofollow">
                     <i class="fa fa-facebook"></i>
@@ -15,7 +15,7 @@
                     <i class="fa fa-facebook"></i>
                 </button>
             @endif
-            @if (!empty($user->profile->twitter))
+            @if ($user->profile->twitter)
                 <a class="btn btn-social-icon btn-twitter" href="{{ $user->profile->twitter }}"
                    rel="nofollow">
                     <i class="fa fa-twitter"></i>
@@ -25,7 +25,7 @@
                     <i class="fa fa-twitter"></i>
                 </button>
             @endif
-            @if (!empty($user->profile->linkedin))
+            @if ($user->profile->linkedin)
                 <a class="btn btn-social-icon btn-linkedin" href="{{ $user->profile->linkedin }}"
                    rel="nofollow">
                     <i class="fa fa-linkedin"></i>
@@ -35,7 +35,7 @@
                     <i class="fa fa-linkedin"></i>
                 </button>
             @endif
-            @if (!empty($user->profile->github))
+            @if ($user->profile->github)
                 <a class="btn btn-social-icon btn-github" href="{{ $user->profile->github }}"
                    rel="nofollow">
                     <i class="fa fa-github"></i>
@@ -50,51 +50,6 @@
         <table class="table table-condensed table-hover">
             <thead>
             <tr>
-                <th colspan="2">@lang('user/profile.contact_info')</th>
-            </tr>
-            </thead>
-            <tbody>
-            <tr>
-                <td>@lang('user/profile.url'):</td>
-                <td>
-                    @if (!empty($user->profile->url))
-                        <a href="{{ $user->profile->url }}" rel="nofollow">{{ $user->profile->url }}</a>
-                    @endif
-                </td>
-
-            </tr>
-            <tr>
-                <td>@lang('user/profile.email'):</td>
-                <td>
-                    <a href="mailto:{{ $user->email }}" rel="nofollow">
-                        {{ $user->email }}
-                    </a>
-                </td>
-            </tr>
-            <tr>
-                <td>@lang('user/profile.phone'):</td>
-                <td>
-                    @if (!empty($user->profile->phone))
-                        <a href="tel:{{ $user->profile->phone }}"
-                           rel="nofollow">{{ $user->profile->phone }}</a>
-                    @endif
-                </td>
-            </tr>
-            <tr>
-                <td>@lang('user/profile.mobile'):</td>
-                <td>
-                    @if (!empty($user->profile->mobile))
-                        <a href="tel:{{ $user->profile->mobile }}"
-                           rel="nofollow">{{ $user->profile->mobile }}</a>
-                    @endif
-                </td>
-            </tr>
-            </tbody>
-        </table>
-        <hr>
-        <table class="table table-condensed table-hover">
-            <thead>
-            <tr>
                 <th colspan="2">@lang('user/profile.general_info')</th>
             </tr>
             </thead>
@@ -104,16 +59,20 @@
                 <td>{{ $user->level }}</td>
             </tr>
             <tr>
+                <td>@lang('user/profile.badges'):</td>
+                <td>{{ $user->unlockedBadgesCount() }}</td>
+            </tr>
+            <tr>
                 <td>@lang('user/profile.experience'):</td>
                 <td>{{ $user->experience }}</td>
             </tr>
             <tr>
                 <td>@lang('user/profile.user_since'):</td>
-                <td>{{ date("M Y", strtotime($user->created_at)) }}</td>
+                <td>{{ $user->present()->createdAt }}</td>
             </tr>
             <tr>
                 <td>@lang('user/profile.roles'):</td>
-                <td>{{ ucfirst($user->role) }}</td>
+                <td>{{ $user->present()->role }}</td>
             </tr>
             </tbody>
         </table>
@@ -126,20 +85,9 @@
             </thead>
             <tbody>
             <tr>
-                <td>@lang('user/profile.birth')</td>
+                <td>@lang('user/profile.date_of_birth')</td>
                 <td>
-                    @if (isset($user->profile->date_of_birth))
-                        {{ date("d F Y", strtotime($user->profile->date_of_birth)) }}
-                    @endif
-                </td>
-            </tr>
-            <tr>
-                <td>@lang('user/profile.gender')</td>
-                <td>
-                    @if (isset($user->profile->gender))
-                        <i class="fa fa-{{ $user->profile->gender }}"></i>
-                        {{ __('user/profile.gender_' . $user->profile->gender) }}
-                    @endif
+                    {{ $user->present()->birthdate }}
                 </td>
             </tr>
             </tbody>