+++ resources/views/profile/_badge_locked.blade.php
@@ -1,0 +1,4 @@
+<div class="col-xs-2">
+        <img src="{{ asset(\Gamify\Models\Badge::DEFAULT_IMAGE) }}" class="img-thumbnail" alt="{{ $badge->name }}"
+             data-toggle="tooltip" data-placement="bottom" title="{{ $badge->name }}">
+</div>