+++ tests/Feature/Http/Controllers/UserControllerTest.php
@@ -25,11 +25,14 @@
 
 namespace Tests\Feature\Http\Controllers;
 
-use Gamify\Enums\Roles;
+use Gamify\Events\UserProfileUpdated;
 use Gamify\Models\User;
+use Gamify\Models\UserProfile;
 use Generator;
 use Illuminate\Foundation\Testing\RefreshDatabase;
-use Illuminate\Support\Facades\Hash;
+use Illuminate\Http\UploadedFile;
+use Illuminate\Support\Facades\Event;
+use Illuminate\Support\Facades\Storage;
 use Tests\TestCase;
 
 class UserControllerTest extends TestCase
@@ -47,7 +50,7 @@ class UserControllerTest extends TestCase
     }
 
     /** @test */
-    public function users_should_see_another_user_profiles(): void
+    public function users_should_see_another_user_profiles_without_edit_buttons(): void
     {
         /** @var User $user */
         $user = User::factory()->create();
@@ -57,496 +60,177 @@ class UserControllerTest extends TestCase
             ->get(route('profiles.show', ['username' => $user->username]))
             ->assertSuccessful()
             ->assertViewIs('profile.show')
-            ->assertViewHas('user', $user);
+            ->assertViewHas('user', $user)
+            ->assertDontSeeText(__('user/profile.edit_account'))
+            ->assertDontSeeText(__('user/profile.change_password'));
     }
 
     /** @test */
-    public function admins_should_see_the_index_view()
+    public function users_should_see_its_own_profile_edit_buttons(): void
     {
-        $this->user->role = Roles::Admin;
-
         $this
             ->actingAs($this->user)
-            ->get(route('admin.users.index'))
+            ->get(route('profiles.show', $this->user->username))
             ->assertSuccessful()
-            ->assertViewIs('admin.user.index');
+            ->assertViewIs('profile.show')
+            ->assertViewHas('user', $this->user)
+            ->assertSeeText(__('user/profile.edit_account'))
+            ->assertSeeText(__('user/profile.change_password'));
     }
 
     /** @test */
-    public function players_should_not_see_the_new_user_form(): void
+    public function users_should_not_update_another_user_profiles(): void
     {
-        $this
-            ->actingAs($this->user)
-            ->get(route('admin.users.create'))
-            ->assertForbidden();
-    }
+        /** @var User $user */
+        $user = User::factory()->create();
 
-    /** @test */
-    public function admins_should_see_the_new_user_form(): void
-    {
-        $this->user->role = Roles::Admin;
+        /** @var UserProfile $want */
+        $want = UserProfile::factory()->make();
 
         $this
             ->actingAs($this->user)
-            ->get(route('admin.users.create'))
-            ->assertSuccessful()
-            ->assertViewIs('admin.user.create');
-    }
-
-    /** @test */
-    public function players_should_not_create_users(): void
-    {
-        /** @var User $want */
-        $want = User::factory()->make();
-
-        $this
-            ->actingAs($this->user)
-            ->post(route('admin.users.store'), [
-                'username' => $want->username,
-                'name' => $want->name,
-                'email' => $want->email,
-                'password' => 'secret123',
-                'password_confirmation' => 'secret123',
-                'role' => $want->role,
+            ->put(route('profiles.update', $user->username), [
+                'bio' => $want->bio,
+                'date_of_birth' => $want->date_of_birth,
+                'twitter' => $want->twitter,
+                'facebook' => $want->facebook,
+                'linkedin' => $want->linkedin,
+                'github' => $want->github,
             ])
             ->assertForbidden();
 
-        $this->assertDatabaseMissing(User::class, [
-            'username' => $want->username,
-            'email' => $want->email,
+        $this->assertDatabaseHas(UserProfile::class, [
+            'user_id' => $user->id,
+            'bio' => $user->profile->bio,
+            'date_of_birth' => $user->profile->date_of_birth,
+            'twitter' => $user->profile->twitter,
+            'facebook' => $user->profile->facebook,
+            'linkedin' => $user->profile->linkedin,
+            'github' => $user->profile->github,
         ]);
     }
 
     /** @test */
-    public function admins_should_create_users(): void
+    public function users_should_update_its_own_profile(): void
     {
-        $this->user->role = Roles::Admin;
-
-        /** @var User $want */
-        $want = User::factory()->make();
+        /** @var UserProfile $want */
+        $want = UserProfile::factory()->make();
 
         $this
             ->actingAs($this->user)
-            ->post(route('admin.users.store'), [
-                'username' => $want->username,
-                'name' => $want->name,
-                'email' => $want->email,
-                'password' => 'secret123',
-                'password_confirmation' => 'secret123',
-                'role' => $want->role->value,
+            ->put(route('profiles.update', $this->user->username), [
+                'bio' => $want->bio,
+                'date_of_birth' => $want->date_of_birth,
+                'twitter' => $want->twitter,
+                'facebook' => $want->facebook,
+                'linkedin' => $want->linkedin,
+                'github' => $want->github,
             ])
-            ->assertRedirect(route('admin.users.index'))
+            ->assertRedirect(route('profiles.show', $this->user->username))
             ->assertValid();
 
-        $this->assertDatabaseHas(User::class, [
-            'username' => $want->username,
-            'name' => $want->name,
-            'email' => $want->email,
-            'role' => $want->role,
+        $this->assertDatabaseHas(UserProfile::class, [
+            'user_id' => $this->user->id,
+            'bio' => $want->bio,
+            'date_of_birth' => $want->date_of_birth,
+            'twitter' => $want->twitter,
+            'facebook' => $want->facebook,
+            'linkedin' => $want->linkedin,
+            'github' => $want->github,
         ]);
     }
 
     /**
      * @test
-     * @dataProvider provideWrongDataForUserCreation
+     * @dataProvider provideWrongDataForUserProfileModification
      */
-    public function admins_should_get_errors_when_creating_users_with_wrong_data(
+    public function users_should_get_errors_when_updating_its_own_profile_with_wrong_data(
         array $data,
         array $errors
     ): void {
-        $this->user->role = Roles::Admin;
+        $userProfile = $this->user->profile;
 
-        // User to validate unique rules...
-        User::factory()->create([
-            'username' => 'john',
-            'email' => 'john.doe@domain.local',
-        ]);
-
-        /** @var User $want */
-        $want = User::factory()->make();
+        /** @var UserProfile $want */
+        $want = UserProfile::factory()->make();
 
         $formData = [
-            'username' => $data['username'] ?? $want->username,
-            'name' => $data['name'] ?? $want->name,
-            'email' => $data['email'] ?? $want->email,
-            'password' => $data['password'] ?? $want->password,
-            'password_confirmation' => $data['password_confirmation'] ?? $want->password,
-            'role' => $data['role'] ?? $want->role->value,
+            'bio' => $data['bio'] ?? $want->bio,
+            'date_of_birth' => $data['date_of_birth'] ?? $want->date_of_birth,
+            'twitter' => $data['twitter'] ?? $want->twitter,
+            'facebook' => $data['facebook'] ?? $want->facebook,
+            'linkedin' => $data['linkedin'] ?? $want->linkedin,
+            'github' => $data['github'] ?? $want->github,
         ];
 
         $this
             ->actingAs($this->user)
-            ->post(route('admin.users.store'), $formData)
+            ->put(route('profiles.update', $this->user->username), $formData)
             ->assertInvalid($errors);
 
-        $this->assertDatabaseMissing(User::class, [
-            'username' => $formData['username'],
-            'email' => $formData['email'],
-        ]);
+        $this->assertModelExists($userProfile);
     }
 
-    public function provideWrongDataForUserCreation(): Generator
+    public function provideWrongDataForUserProfileModification(): Generator
     {
-        yield 'username is empty' => [
+        yield 'birthdate ! a date' => [
             'data' => [
-                'username' => '',
+                'date_of_birth' => 'foo',
             ],
-            'errors' => ['username'],
-        ];
-
-        yield 'username > 255 chars' => [
-            'data' => [
-                'username' => '01234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012
-34567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345',
-            ],
-            'errors' => ['username'],
-        ];
-
-        yield 'username ! a username' => [
-            'data' => [
-                'username' => 'u$ern4me',
-            ],
-            'errors' => ['username'],
-        ];
-
-        yield 'username is taken' => [
-            'data' => [
-                'username' => 'john',
-            ],
-            'errors' => ['username'],
-        ];
-
-        yield 'name is empty' => [
-            'data' => [
-                'name' => '',
-            ],
-            'errors' => ['name'],
-        ];
-
-        yield 'email is empty' => [
-            'data' => [
-                'email' => '',
-            ],
-            'errors' => ['email'],
-        ];
-
-        yield 'email ! an email' => [
-            'data' => [
-                'email' => 'is-not-an-email',
-            ],
-            'errors' => ['email'],
-        ];
-
-        yield 'email is taken' => [
-            'data' => [
-                'email' => 'john.doe@domain.local',
-            ],
-            'errors' => ['email'],
-        ];
-
-        yield 'password is empty' => [
-            'data' => [
-                'password' => '',
-            ],
-            'errors' => ['password'],
-        ];
-
-        yield 'password ! long enough' => [
-            'data' => [
-                'password' => '1234',
-            ],
-            'errors' => ['password'],
-        ];
-
-        yield 'password ! confirmed' => [
-            'data' => [
-                'password' => 'verySecretPassword',
-                'password_confirmation' => 'notSoSecretPassword',
-            ],
-            'errors' => ['password'],
-        ];
-
-        yield 'role ! a role' => [
-            'data' => [
-                'role' => 'non-existent-role',
-            ],
-            'errors' => ['role'],
+            'errors' => ['date_of_birth'],
         ];
     }
 
     /** @test */
-    public function players_should_not_see_any_user(): void
+    public function users_should_update_its_own_avatar(): void
     {
-        $want = User::factory()->create();
+        Storage::fake('public');
+
+        $file = UploadedFile::fake()->image('avatar.jpg');
 
         $this
             ->actingAs($this->user)
-            ->get(route('admin.users.show', $want))
-            ->assertForbidden();
-    }
-
-    /** @test */
-    public function admins_should_see_any_user(): void
-    {
-        $this->user->role = Roles::Admin;
-
-        $want = User::factory()->create();
-
-        $this
-            ->actingAs($this->user)
-            ->get(route('admin.users.show', $want))
-            ->assertSuccessful()
-            ->assertViewIs('admin.user.show')
-            ->assertViewHas('user', $want);
-    }
-
-    /** @test */
-    public function players_should_not_see_the_edit_user_form(): void
-    {
-        $want = User::factory()->create();
-
-        $this
-            ->actingAs($this->user)
-            ->get(route('admin.users.edit', $want))
-            ->assertForbidden();
-    }
-
-    /** @test */
-    public function admins_should_see_the_edit_user_form(): void
-    {
-        $this->user->role = Roles::Admin;
-
-        $want = User::factory()->create();
-
-        $this
-            ->actingAs($this->user)
-            ->get(route('admin.users.edit', $want))
-            ->assertSuccessful()
-            ->assertViewIs('admin.user.edit')
-            ->assertViewHas('user', $want);
-    }
-
-    /** @test */
-    public function players_should_not_update_users(): void
-    {
-        $want = User::factory()->create();
-
-        $this
-            ->actingAs($this->user)
-            ->put(route('admin.users.update', $want), [])
-            ->assertForbidden();
-
-        $this->assertModelExists($want);
-    }
-
-    /** @test */
-    public function admins_should_update_users(): void
-    {
-        $this->user->role = Roles::Admin;
-
-        /** @var User $user */
-        $user = User::factory()->create();
-
-        /** @var User $want */
-        $want = User::factory()->make();
-
-        $this
-            ->actingAs($this->user)
-            ->put(route('admin.users.update', $user), [
-                'name' => $want->name,
-                'email' => $want->email,
-                'password' => 'new-password-123',
-                'password_confirmation' => 'new-password-123',
-                'role' => $want->role->value,
+            ->put(route('profiles.update', $this->user->username), [
+                'image' => $file,
             ])
-            ->assertRedirect(route('admin.users.edit', $user))
+            ->assertRedirect(route('profiles.show', $this->user->username))
             ->assertValid();
 
-        $this->assertDatabaseHas(User::class, [
-            'id' => $user->id,
-            'name' => $want->name,
-            'email' => $want->email,
-            'role' => $want->role,
-        ]);
+        Storage::disk('public')->assertExists('avatars/' . $file->hashName());
 
-        $user->refresh();
-
-        $this->assertTrue(Hash::check('new-password-123', $user->password));
+        $this->assertEquals(Storage::url('avatars/' . $file->hashName()), $this->user->profile->avatarUrl);
     }
 
     /** @test */
-    public function admins_should_not_update_their_own_role(): void
+    public function event_should_be_dispatched_when_user_profile_is_updating_at_least_one_attribute(): void
     {
-        $this->user->role = Roles::Admin;
-
-        $this->user->save();
-
-        /** @var User $want */
-        $want = User::factory()->make();
+        Event::fake([
+            UserProfileUpdated::class,
+        ]);
 
         $this
             ->actingAs($this->user)
-            ->put(route('admin.users.update', $this->user), [
-                'name' => $want->name,
-                'email' => $want->email,
-                'role' => Roles::Player,
+            ->put(route('profiles.update', $this->user->username), [
+                'bio' => 'foo',
             ])
-            ->assertRedirect(route('admin.users.edit', $this->user))
+            ->assertRedirect(route('profiles.show', $this->user->username))
             ->assertValid();
 
-        $this->assertDatabaseHas(User::class, [
-            'username' => $this->user->username,
-            'name' => $want->name,
-            'email' => $want->email,
-            'role' => $this->user->role,
-        ]);
-    }
-
-    /**
-     * @test
-     * @dataProvider provideWrongDataForUserModification
-     */
-    public function admins_should_get_errors_when_updating_users_with_wrong_data(
-        array $data,
-        array $errors
-    ): void {
-        $this->user->role = Roles::Admin;
-
-        // User to validate unique rules...
-        User::factory()->create([
-            'username' => 'john',
-            'email' => 'john.doe@domain.local',
-        ]);
-
-        /** @var User $user */
-        $user = User::factory()->create([
-            'password' => 'veryS3cr3t',
-        ]);
-
-        $formData = [
-            'name' => $data['name'] ?? $user->name,
-            'email' => $data['email'] ?? $user->email,
-            'password' => $data['password'] ?? $user->password,
-            'password_confirmation' => $data['password_confirmation'] ?? $user->password,
-            'role' => $data['role'] ?? $user->role->value,
-        ];
-
-        $this
-            ->actingAs($this->user)
-            ->put(route('admin.users.update', $user), $formData)
-            ->assertInvalid($errors);
-
-        $this->assertDatabaseHas(User::class, [
-            'id' => $user->id,
-            'username' => $user->username,
-            'name' => $user->name,
-            'email' => $user->email,
-            'role' => $user->role,
-        ]);
-
-        $user->refresh();
-
-        $this->assertTrue(Hash::check('veryS3cr3t', $user->password));
-    }
-
-    public function provideWrongDataForUserModification(): Generator
-    {
-        yield 'name is empty' => [
-            'data' => [
-                'name' => '',
-            ],
-            'errors' => ['name'],
-        ];
-
-        yield 'email is empty' => [
-            'data' => [
-                'email' => '',
-            ],
-            'errors' => ['email'],
-        ];
-
-        yield 'email ! an email' => [
-            'data' => [
-                'email' => 'is-not-an-email',
-            ],
-            'errors' => ['email'],
-        ];
-
-        yield 'email is taken' => [
-            'data' => [
-                'email' => 'john.doe@domain.local',
-            ],
-            'errors' => ['email'],
-        ];
-
-        yield 'password is empty' => [
-            'data' => [
-                'password' => '',
-            ],
-            'errors' => ['password'],
-        ];
-
-        yield 'password ! long enough' => [
-            'data' => [
-                'password' => '1234',
-            ],
-            'errors' => ['password'],
-        ];
-
-        yield 'password ! confirmed' => [
-            'data' => [
-                'password' => 'verySecretPassword',
-                'password_confirmation' => 'notSoSecretPassword',
-            ],
-            'errors' => ['password'],
-        ];
-
-        yield 'role ! a role' => [
-            'data' => [
-                'role' => 'non-existent-role',
-            ],
-            'errors' => ['role'],
-        ];
+        Event::assertDispatched(UserProfileUpdated::class);
     }
 
     /** @test */
-    public function players_should_not_delete_users(): void
+    public function event_should_not_be_dispatched_when_user_profile_is_not_updating_any_attribute(): void
     {
-        /** @var User $user */
-        $user = User::factory()->create();
+        Event::fake([
+            UserProfileUpdated::class,
+        ]);
 
         $this
             ->actingAs($this->user)
-            ->get(route('admin.users.destroy', $user))
-            ->assertForbidden();
-    }
+            ->put(route('profiles.update', $this->user->username), [])
+            ->assertRedirect(route('profiles.show', $this->user->username))
+            ->assertValid();
 
-    /** @test */
-    public function admins_should_not_delete_themselves(): void
-    {
-        $this->user->role = Roles::Admin;
-
-        $this
-            ->actingAs($this->user)
-            ->delete(route('admin.users.destroy', $this->user))
-            ->assertForbidden();
-
-        $this->assertModelExists($this->user);
-    }
-
-    /** @test */
-    public function admins_should_delete_users(): void
-    {
-        $this->user->role = Roles::Admin;
-
-        /** @var User $user */
-        $user = User::factory()->create();
-
-        $this
-            ->actingAs($this->user)
-            ->delete(route('admin.users.destroy', $user))
-            ->assertRedirect(route('admin.users.index'));
-
-        $this->assertModelMissing($user);
+        Event::assertNotDispatched(UserProfileUpdated::class);
     }
 }