+++ tests/Feature/Listeners/IncrementBadgeOnQuestionAnsweredTest.php
@@ -72,7 +72,7 @@ class IncrementBadgeOnQuestionAnsweredTest extends TestCase
         $listener = new IncrementBadgesOnQuestionAnswered();
         $listener->handle($event);
 
-        $this->assertEquals(1, $user->badges()->wherePivot('badge_id', $badge->id)->first()->pivot->repetitions);
+        $this->assertEquals(1, $user->badges()->wherePivot('badge_id', $badge->id)->first()->progress->repetitions);
     }
 
     /** @test */
@@ -122,7 +122,7 @@ class IncrementBadgeOnQuestionAnsweredTest extends TestCase
         $listener = new IncrementBadgesOnQuestionAnswered();
         $listener->handle($event);
 
-        $this->assertEquals(1, $user->badges()->wherePivot('badge_id', $badge->id)->first()->pivot->repetitions);
+        $this->assertEquals(1, $user->badges()->wherePivot('badge_id', $badge->id)->first()->progress->repetitions);
     }
 
     /** @test */
@@ -147,7 +147,7 @@ class IncrementBadgeOnQuestionAnsweredTest extends TestCase
         $listener = new IncrementBadgesOnQuestionAnswered();
         $listener->handle($event);
 
-        $this->assertEquals(1, $user->badges()->wherePivot('badge_id', $badge->id)->first()->pivot->repetitions);
+        $this->assertEquals(1, $user->badges()->wherePivot('badge_id', $badge->id)->first()->progress->repetitions);
     }
 
     /** @test */